### 2021-01-04 #2936 @whedon
[PRE REVIEW]: MOLE: Mimetic Operators Library Enhanced
**Submitting author:** @jcorbino (<a href="http://orcid.org/0000-0002-2638-9216">Johnny Corbino</a>)
**Repository:** <a href="https://github.com/jcorbino/mole" target ="_blank">https://github.com/jcorbino/mole</a>
**Version:** v3.0.0
