### 2021-01-07 #2936 @jedbrown
Hi, @jcorbino. :wave:  At first glance, I'm curious if the documentation is rendered somewhere and how one goes about testing the software. The documentation looks to be a mix of PDF files of unknown origin and some generated Doxygen docs that have been committed to the repository. It's quite easy these days to use an action to build a GitHub Pages site containing rendered documentation. And there should be source files for all content so it can be edited -- I see the Users Manual only describes the MATLAB interface and has outdated citations.

Are there external research users of the software? If so, what do they look like?

The C++ code size is small enough that we tend to ask questions: https://joss.readthedocs.io/en/latest/submitting.html?highlight=maintainable#substantial-scholarly-effort
I see the MATLAB is about 2000 lines, but it's unclear to me how the capabilities compare. Much of the question here is whether the software is something that the research community builds upon, or whether it's more of a reproducibility pack for your 2020 paper that readers of the paper might copy and modify for their use.

> JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools).

Could you please update the repository to address these issues or ask questions here in this issue?
