### 2021-01-07 #2936 @jcorbino
Hi @jedbrown, thank you for your feedback. This is basically what we have (documentation-wise):

**#1** _mole/doc_C++_  is our Doxygen generated documentation for the C++ version of the library
**#2** _mole/doc_MATLAB_ is the API documentation for MOLE's MATLAB version (generated through m2html)
**#3** _mole/MOLE User's Manual (MATLAB version).pdf_ is a short document that explains the resolution of a common problem using MOLE (yes, as you mentioned, only the MATLAB version)
**#4** There are a couple of extra PDFs that just talk about why we use staggered grids etc.
**#5** There are two examples folders (one for each flavor)

It is important to mention that the MATLAB version of our library is ahead of the C++ version, that is, there are certain functions (e.g. curvilinear operators) that so far have only been implemented for MATLAB.

So far I'm the sole contributor of the library (I mention this just to clarify on one of the rules of JOSS regarding the minimum of 3 months work for individual repos).

The library is backed by a peer-reviewed paper published (as you clearly mentioned), in addition, a couple of articles have been published from students in our lab that have used MOLE for their research [(](https://www.sciencedirect.com/science/article/pii/S0045793020303169?via%3Dihub)  and [)](https://ojs.lib.uwo.ca/index.php/mase/article/view/10874) there are other articles but I don't see the point in mentioning them.

Regarding the "_feature-complete, half-baked solutions_" I'm a bit puzzled, as a developer, I do not see how the current state of the library does not fulfill such requirement. The library allows to solve steady-state and time-dependent problems using mimetic methods in 1, 2 and 3 spatial dimensions. The reproducibility of our paper is just a subset of the set of applications.

Yes, there are internal and external users of the library, most of them graduate students. Most of them are Computer Scientists, Mathematicians, and Physicists.

Regarding the GitHub Page, so far I just have [](https://jcorbino.github.io/mole/) as the "page". I know it would be ideal to embed all or part of the documentation into that page but right now that is not a priority.
