### 2021-01-09 #2936 @jcorbino
@jedbrown "User's manual" has been updated. Thanks for pointing that out!

Regarding having a more detailed GitHub page (so the different elements of documentation can be accessed from an index.md), currently I'm getting myself familiar with the creation process (I think it is a good addition), but as I mentioned in my previous comment, this has low priority.
