### 2021-01-25 #2936 @rouson
@jedbrown  I just gave @jcorbino several tips to help get the MOLE repository in shape to meet the JOSS review criteria.  I also suggested that the JOSS paper.md write-up be revised so that it cites the research literature in which the submitted code has been used.  Hopefully this will demonstrate the substantial scholarly contributions of the code.
