### 2021-01-27 9c7d442
Updated installation instructions
