### 2021-01-27 #2936 @jcorbino
Thanks @rouson for your feedback. @jedbrown , I've updated the following:

- *doc_MATLAB* (generated documentation about the MATLAB implementation of MOLE)
- *README.md* (added some extra details to make installation process clear)
- Updated references
