### 2021-02-10 #2936 @jcorbino
Hi @jedbrown , I totally understand, I'm in a similar situation, kids + COVID restrictions make the system nonlinear. **I would like to proceed with the manuscript in this journal**, but I don't know exactly how to do it.  For reviewing I think that david.ketcheson@kaust.edu.sa (David Ketcheson) and damian@sourceryinstitute.org (Damian Rouson) are appropriate (based on their area of expertise).  @arfon , is there any way to tell the system to reverse the withdrawal action?  Thanks!
