### 2021-02-14 ac96325
New 2D test case. Solution to NS equations using mimetics
It solves two-dimensional Navier-Stokes equation with Boussinesq's approximation via mimetic methods. For sharp Kelvin-Helmholtz billows, use a dx == 0.0625. The most time-consuming part (as expected) is the computation of the pressure field, Laplacian matrix is highly sparse but not positive definite.
