+++ examples_MATLAB/lock_exchange.m
@@ -1,0 +1,269 @@
+%% 2D Lock exchange test case using mimetic methods
+% It uses Boussinesq's approximation to solve for nonisothermal flow.
+% Due to the coarseness of the grid being used, numerical diffusion acts to
+% limit the formation of the billows along the interface. Sharp billows can
+% easily be obtained by increasing the resolution.
+
+clc
+close all
+
+%% MOLE'S path so mimetic methods can be used
+addpath('mole/mole_MATLAB')
+
+tic  % Start the timer
+%--------------------------------------------------------------------------
+
+%% Grid [a b]x[c d] in meters
+a = 0;
+b = 100;
+c = 0;
+d = 20;
+m = 100;       % Number of cells along x-axis
+n = 20;        % Number of cells along y-axis
+dx = (b-a)/m;  % Step size along x- and y-axis, respectively
+dy = (d-c)/n;
+
+% Rectangular grid with physical coordinates
+[X, Y] = meshgrid([a a+dx/2 : dx : b-dx/2 b], [c c+dy/2 : dy : d-dy/2 d]);
+
+% Density
+rho = zeros(n+2, m+2);
+% Temperature
+T = zeros(n+2, m+2);
+% Pressure
+p = zeros(n+2, m+2);
+% u-velocity
+u = zeros(n, m+1);
+% v-velocity
+v = zeros(n+1, m);
+
+%% Parameters (for seawater @ S = 35)
+alpha = 1.664e-4;        % 1/°C Coefficient of thermal expansion
+T_0 = 10;                % °C Reference temperature
+rho_0 = 1027;            % kg/m3 Reference density
+interface_width = 0.1;   % m Width of the interface
+g = 9.806;               % m/s2 Gravity acceleration
+reduced_gravity = 0.01;  % m/s2 Reduced gravity
+delta_rho = reduced_gravity*rho_0/g;  % Density difference between the two fluids
+
+% Time parameters
+time = 200;              % s Simulation time
+dt = 1;                  % s Time step
+iterations = time/dt;    % Number of iterations
+
+%% Initialize densities and temperatures
+for j = 1 : m+2
+    x = a + (j-1)*dx;
+    rho(:, j) = rho_0+delta_rho/2*(1-erf((x-(b-a)/2)/interface_width));
+    T(:, j) = (1-rho(1, j)/rho_0)/alpha+T_0;
+end
+
+%% Middle points and viscosities
+T_middle = (T(1)+T(end))/2;
+rho_middle = rho_0*(1-alpha*(T_middle-T_0));  % Could also be calculated as simply (rho(1)+rho(end))/2
+mu = 0.00141;                 % kg/(m⋅s) Dynamic viscosity
+nu = mu/rho_middle;           % m2/s Kinematic viscosity
+
+%% Mimetic operators
+k = 2;                        % Spatial order of accuracy
+D = div2D(k, m, dx, n, dy);   % Divergence
+G = grad2D(k, m, dx, n, dy);  % Gradient
+L = D*G;                      % Laplacian
+
+% Premultiplication of G
+G = -dt/rho_middle*G;
+
+% Apply Neumann BCs to Laplacian
+L = L+robinBC2D(k, m, dx, n, dy, 0, 1);
+
+% Interpolators
+I0 = interpol2D(m, n, 0, 0);  % For downwind
+I1 = interpol2D(m, n, 1, 1);  % For upwind
+
+%% Iterate over time
+SOL = zeros(size(D, 2), 1);   % To be used later for heat advection
[…]
