+++ README.md
@@ -65,6 +65,7 @@ the library https://www.sciencedirect.com/science/article/abs/pii/S0377042719303
 
 [![View mole on File Exchange](https://www.mathworks.com/matlabcentral/images/matlab-file-exchange.svg)](https://www.mathworks.com/matlabcentral/fileexchange/64095-mole)
 
+![Obtained with curvilinear operators](images/200s_high_resolution_hot_edges.png)
 ![Obtained with curvilinear operators](images/4thOrder.png)
 ![Obtained with curvilinear operators](images/4thOrder2.png)
 ![Obtained with curvilinear operators](images/4thOrder3.png)
