### 2021-02-15 387f474
Minor optimization of the code
