### 2021-02-15 1643da8
Update lock_exchange.m
