### 2021-02-15 fad648f
Added comment about CFL condition
