### 2021-02-21 9247d32
Delete doc_C++ directory
