+++ doc_C++/search/classes_1.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['gradient',['Gradient',['../classGradient.html',1,'']]]
-];
