+++ doc_C++/search/files_0.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['divergence_2ecpp',['divergence.cpp',['../divergence_8cpp.html',1,'']]],
-  ['divergence_2eh',['divergence.h',['../divergence_8h.html',1,'']]]
-];
