+++ doc_C++/search/functions_2.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['interpol',['Interpol',['../classInterpol.html#a3b7e86f50c1140e0645bbc5380cbe151',1,'Interpol::Interpol(u32 m, double c)'],['../classInterpol.html#ab2918bea6651d9c37dca00c538964d96',1,'Interpol::Interpol(u32 m, u32 n, double c1, double c2)'],['../classInterpol.html#a2bcc732c39bcfd7ae2435c2119e5f560',1,'Interpol::Interpol(u32 m, u32 n, u32 o, double c1, double c2, double c3)']]]
-];
