+++ doc_C++/search/functions_0.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['divergence',['Divergence',['../classDivergence.html#ac714d81fb8f0b8f601f07115dd514380',1,'Divergence::Divergence(u16 k, u32 m, double dx)'],['../classDivergence.html#a993e5888f9600ec83c350ef3fd8d08ed',1,'Divergence::Divergence(u16 k, u32 m, u32 n, double dx, double dy)'],['../classDivergence.html#a036305435cef8b28198563ce292f7ed1',1,'Divergence::Divergence(u16 k, u32 m, u32 n, u32 o, double dx, double dy, double dz)']]]
-];
