+++ doc_C++/search/functions_4.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['operator_2a',['operator*',['../operators_8h.html#af587420d81213bb9cf4bdceba8adaec5',1,'operator*(const Divergence &amp;div, const Gradient &amp;grad):&#160;operators.h'],['../operators_8h.html#aa0f3c6d9ffa98e3308f1c654d7cb0d07',1,'operator*(const Divergence &amp;div, const vec &amp;v):&#160;operators.h'],['../operators_8h.html#a02455656721dcb2948197aa6c053313a',1,'operator*(const Gradient &amp;grad, const vec &amp;v):&#160;operators.h'],['../operators_8h.html#a221ecf61e403f612c8a4811c88038b42',1,'operator*(const Laplacian &amp;lap, const vec &amp;v):&#160;operators.h'],['../operators_8h.html#a25e80adf88a4f071196e7ca1920cd164',1,'operator*(const Interpol &amp;I, const vec &amp;v):&#160;operators.h']]]
-];
