+++ doc_C++/search/files_3.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['laplacian_2ecpp',['laplacian.cpp',['../laplacian_8cpp.html',1,'']]],
-  ['laplacian_2eh',['laplacian.h',['../laplacian_8h.html',1,'']]]
-];
