+++ doc_C++/search/files_2.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['interpol_2ecpp',['interpol.cpp',['../interpol_8cpp.html',1,'']]],
-  ['interpol_2eh',['interpol.h',['../interpol_8h.html',1,'']]]
-];
