+++ doc_C++/search/classes_2.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['interpol',['Interpol',['../classInterpol.html',1,'']]]
-];
