+++ doc_C++/search/all_1.js
@@ -1,8 +1,0 @@
-var searchData=
-[
-  ['getp',['getP',['../classGradient.html#a6099174246fa7784d9702b5cf26694d7',1,'Gradient']]],
-  ['getq',['getQ',['../classDivergence.html#a9ab0da6ee16064012c164592648e66c5',1,'Divergence']]],
-  ['gradient',['Gradient',['../classGradient.html',1,'Gradient'],['../classGradient.html#a74a11d8acd2b6aef856cd47aef77283b',1,'Gradient::Gradient(u16 k, u32 m, double dx)'],['../classGradient.html#a3aeff5685f81c30812a7749a6267942d',1,'Gradient::Gradient(u16 k, u32 m, u32 n, double dx, double dy)'],['../classGradient.html#a8b868a729f9e9078733a3ec6ac143481',1,'Gradient::Gradient(u16 k, u32 m, u32 n, u32 o, double dx, double dy, double dz)']]],
-  ['gradient_2ecpp',['gradient.cpp',['../gradient_8cpp.html',1,'']]],
-  ['gradient_2eh',['gradient.h',['../gradient_8h.html',1,'']]]
-];
