+++ doc_C++/search/functions_3.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['laplacian',['Laplacian',['../classLaplacian.html#ad7c384323a6042709fbbeea9f806fc2c',1,'Laplacian::Laplacian(u16 k, u32 m, double dx)'],['../classLaplacian.html#a5a4e50cc5e4a37200b561d128afa2888',1,'Laplacian::Laplacian(u16 k, u32 m, u32 n, double dx, double dy)'],['../classLaplacian.html#a1f08dfabf624517209a10308751b1eaa',1,'Laplacian::Laplacian(u16 k, u32 m, u32 n, u32 o, double dx, double dy, double dz)']]]
-];
