+++ .github/workflows/draft-pdf.yml
@@ -8,7 +8,7 @@ jobs:
       - name: Build documentation
         uses: mattnotmitt/doxygen-action@v1
         with:
-          doxyfile-path: ./Doxyfile
+          doxyfile-path: ../../Doxyfile
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
