+++ .github/workflows/draft-pdf.yml
@@ -4,11 +4,12 @@ jobs:
   doxygen:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v1
+      - name: Checkout
+      - uses: actions/checkout@v2
       - name: Build documentation
         uses: mattnotmitt/doxygen-action@v1
         with:
-          doxyfile-path: ../../Doxyfile
+          doxyfile-path: Doxyfile
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
