+++ .github/workflows/draft-pdf.yml
@@ -5,7 +5,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-      - uses: actions/checkout@v2
+        uses: actions/checkout@v2
       - name: Build documentation
         uses: mattnotmitt/doxygen-action@v1
         with:
