+++ .github/workflows/draft-pdf.yml
@@ -1,15 +1,6 @@
 on: [push]
 
 jobs:
-  doxygen:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Build documentation
-        uses: mattnotmitt/doxygen-action@v1
-        with:
-          doxyfile-path: Doxyfile
   paper:
     runs-on: ubuntu-latest
     name: Paper Draft
