+++ .github/workflows/draft-pdf.yml
@@ -14,10 +14,10 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: JOSS/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v2
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: JOSS/paper.pdf
+          path: JOSS/paper.zip
