+++ .github/workflows/draft-pdf.yml
@@ -11,13 +11,9 @@ jobs:
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
-          # This should be the path to the paper within your repo.
           paper-path: JOSS/paper.md
-      - name: Upload
+      - name: Upload a Build Artifact
         uses: actions/upload-artifact@v2
         with:
           name: paper
-          # This is the output path where Pandoc will write the compiled
-          # PDF. Note, this should be the same directory as the input
-          # paper.md
-          path: paper.pdf
+          path: JOSS/paper.pdf
