+++ .github/workflows/doxygen.yml
@@ -9,3 +9,8 @@ jobs:
         uses: mattnotmitt/doxygen-action@v1
         with:
           doxyfile-path: ./Doxyfile
+      - name: Deploy documentation
+        uses: peaceiris/actions-gh-pages@v3
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          publish_dir: ./doc_C++
