+++ search/searchdata.js
@@ -1,0 +1,27 @@
+var indexSectionsWithContent =
+{
+  0: "dgilmnorsu",
+  1: "dgilru",
+  2: "dgilmoru",
+  3: "dgilors",
+  4: "n"
+};
+
+var indexSectionNames =
+{
+  0: "all",
+  1: "classes",
+  2: "files",
+  3: "functions",
+  4: "pages"
+};
+
+var indexSectionLabels =
+{
+  0: "All",
+  1: "Classes",
+  2: "Files",
+  3: "Functions",
+  4: "Pages"
+};
+
