+++ search/functions_5.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['robinbc_54',['RobinBC',['../classRobinBC.html#a17c0fe9978db1e0054da9233e69d8e8b',1,'RobinBC::RobinBC(u16 k, u32 m, double dx, double a, double b)'],['../classRobinBC.html#ac7d4153cc14c34081e2fd8208d8fde49',1,'RobinBC::RobinBC(u16 k, u32 m, double dx, u32 n, double dy, double a, double b)'],['../classRobinBC.html#a3083233aaa22ac3993f5ced010fc70f1',1,'RobinBC::RobinBC(u16 k, u32 m, double dx, u32 n, double dy, u32 o, double dz, double a, double b)']]]
+];
