+++ search/functions_6.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['spjoin_5fcols_55',['spjoin_cols',['../classUtils.html#a802f3347143135fbca199f9a968c8fe5',1,'Utils']]],
+  ['spjoin_5frows_56',['spjoin_rows',['../classUtils.html#a642bdd811b8436603d7153c88c9aec4e',1,'Utils']]],
+  ['spkron_57',['spkron',['../classUtils.html#a24c8eddb344686baa0964b3b5f9a6d51',1,'Utils']]]
+];

[…]
