+++ .github/workflows/doxygen.yml
@@ -1,16 +1,0 @@
-on: [push]
-
-jobs:
-  doxygen:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build documentation
-        uses: mattnotmitt/doxygen-action@v1
-        with:
-          doxyfile-path: ./Doxyfile
-      - name: Deploy documentation
-        uses: peaceiris/actions-gh-pages@v3
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./doc_C++
