### 2021-02-21 #2936 @jcorbino
- Updated the Doxyfile
- Removed generated files (doc_C++ folder)
- Left in place 'doc_MATLAB' folder since this is generated through a non-standard tool
- Expanded a little bit the README file
- Installation looks straight forward
- All files pertinent to JOSS are located in JOSS/
- Created a .github/workflow so paper.pdf gets generated as an artifact on every push

Currently working on catching up the C++ implementation...
