### 2021-02-21 d92714a
Update elliptic1D.cpp
