### 2021-02-22 2fc5f8f
Update burgers1D.m
