### 2021-06-17 #2936 @arfon
@jedbrown Whedon is failing here as it looks like the author may have withdrawn the submission on the JOSS site.

@jcorbino - can you confirm you're still interested in pursuing this publication in JOSS?
