### 2021-06-17 #2936 @danielskatz
So it seems like this should go forward, but that perhaps there is (was) an issue with how whedon handled the author's request to withdraw, which it should have ignored, or maybe something else went wrong along the way
