### 2021-06-17 #2936 @jcorbino
Hi @arfon ,

Yes, I'm still interested in pursing this publication in JOSS. And as @danielskatz pointed out, at some point in time I tried to withdraw the article but it kind of ended up in limbo. If there's anything I can do to fix this, please let me know.
