### 2021-06-17 #2936 @danielskatz
I'm fairly sure this can't be fixed via a whedon command, and my guess is that @openjournals/dev can fix it rather than us having to start it over
