### 2021-06-18 #3380 @whedon
[REVIEW]: MOLE: Mimetic Operators Library Enhanced
**Submitting author:** @jcorbino (<a href="http://orcid.org/0000-0002-2638-9216">Johnny Corbino</a>)
**Repository:** https://github.com/jcorbino/mole
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.0
