### 2021-06-18 #3380 @jedbrown
@ketch @jshipton :wave: Welcome to JOSS and thanks for agreeing to review!

The comments from @whedon up top :point_up: outline the review process, which takes place in this thread (possibly with issues filed in the MOLE repository). I'll be watching this thread if you have any questions.

We typically aim for reviews to be completed within a month or so. (I know you have constraints and may have limited time until August. If you have initial reactions, feel free to share them early so they can be addressed before more detailed review.)  Please let me know if you have further questions. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
