### 2021-06-22 7e0cbbd
Update test_grad3DCurv.m
