### 2021-06-28 bc7ab6b
Create test_lap3DCurv.m
