### 2021-07-03 c817d68
Create amean.m
