+++ mole_MATLAB/amean.m
@@ -1,0 +1,9 @@
+function Y = amean(X)
+% Returns the arithmetic mean for every two pairs of a column vector
+% And, Y(1) = X(1), Y(end) = X(end)
+%
+% Parameters:
+%                X : Column vector
+
+    Y = [X(1); (X(1:end-1)+X(2:end))/2; X(end)];
+end
