### 2021-07-03 7d89487
Create hmean.m
