### 2021-07-03 7ff5b47
Update amean.m
