### 2021-07-03 c297eac
Update hmean.m
