### 2021-07-06 7dbcb94
Added missing flag for the linker
