### 2021-07-06 91edd6e
Update RK4.m
This example is simply to illustrate the implementation of a traditional Runge-Kutta scheme. It doesn't involve mimetic operators.
