### 2021-07-07 05e4d0b
Update RK4.m
This example is simply to illustrate the implementation of a traditional Runge-Kutta scheme. It doesn't involve mimetic operators.
