### 2021-07-07 3f57849
Create RK2.m
This example is simply to illustrate the implementation of RK2. It doesn't involve mimetic operators.
