### 2021-07-07 6ab18b4
Create rk4.m
Explicit Runge-Kutta 4 method for time integration.
