+++ mole_MATLAB/rk4.m
@@ -1,0 +1,24 @@
+function [t, y] = rk4(func, tspan, dt, y0)
+% Explicit Runge-Kutta 4th-order method
+%
+% Returns: t (evaluation points) and y (solutions) of the specified ODE
+%
+% Parameters:
+%                func : Function handler
+%               tspan : [t0 tf]
+%                  dt : Step size
+%                  y0 : Initial conditions
+
+    t = tspan(1) : dt : tspan(2);
+    y = zeros(length(y0), length(t));
+    y(:, 1) = y0;
+    
+    for i = 1 : length(t) - 1
+        k1 = func(t(i),        y(:, i));
+        k2 = func(t(i) + dt/2, y(:, i) + dt/2*k1);
+        k3 = func(t(i) + dt/2, y(:, i) + dt/2*k2);
+        k4 = func(t(i) + dt,   y(:, i) + dt*k3);
+        
+        y(:, i + 1) = y(:, i) + dt/6*(k1 + 2*k2 + 2*k3 + k4);
+    end
+end
