### 2021-07-19 8e51567
Update test_curl.m
