+++ examples_MATLAB/test_curl.m
@@ -56,8 +56,6 @@ quiver3(X(2:end, 2:end), Y(2:end, 2:end), zeros(m, n),...
         U(2:end, 2:end), V(2:end, 2:end), curlMOLE,... 
         'AutoScale', 'off');
 title('Quiver plot');
-set(gca, 'XDir', 'reverse')
-set(gca, 'YDir', 'reverse')
 xlabel('x')
 ylabel('y')
 zlabel('z')
