### 2021-07-19 97e30f4
Deleted view angle
