### 2021-07-19 640c016
Update test_curl.m
