### 2021-07-19 dfc41ba
Update test_curl.m
