### 2021-07-19 d6d7e40
Update test_curl.m
