### 2021-07-19 f61cf30
Update test_curl.m
