### 2021-07-19 e048d94
Update test_curl.m
