+++ examples_MATLAB/test_curl.m
@@ -22,37 +22,33 @@ xaxis = west : (dx/2) : east;
 yaxis = south : (dy/2) : north;
 [X, Y] = meshgrid(xaxis(1:2:end), yaxis(1:2:end));
 
-A1 = zeros(2*m*n+m+n, 1);
-A2 = zeros(2*m*n+m+n, 1);
-
-% Vector field
-U = P(X, Y);
-V = Q(X, Y);
+F = zeros(2*m*n+m+n, 1);
 
 k = 1;
 for j = 2 : 2 : 2*n+1
     for i = 1 : 2 : 2*m+1
-        A1(k) = P(xaxis(i), yaxis(j));
-        A2(k) = Q(xaxis(i), yaxis(j));
+        F(k) = Q(xaxis(i), yaxis(j)); % Important!
         k = k + 1;
     end
 end
 
 for j = 1 : 2 : 2*n+1
     for i = 2 : 2 : 2*m+1
-        A1(k) =  Q(xaxis(i), yaxis(j));
-        A2(k) = -P(xaxis(i), yaxis(j));
+        F(k) = -P(xaxis(i), yaxis(j)); % Important!
         k = k + 1;   
     end
 end
 
-curl = div2D(order, m, dx, n, dy)*A2;
+curl = div2D(order, m, dx, n, dy)*F; % F is F* in https://doi.org/10.1016/j.cam.2019.06.042
 curl = reshape(curl, m+2, n+2);
 curl = curl(2:end-1, 2:end-1);
 
-quiver3(X(2:end, 2:end), Y(2:end, 2:end), zeros(m, n),...
-        U(2:end, 2:end), V(2:end, 2:end), curl,...
-        'AutoScale', 'on');
+% Vector field
+U = P(X, Y);
+V = Q(X, Y);
+
+quiver3(X(2:end, 2:end), Y(2:end, 2:end), zeros(m, n), U(2:end, 2:end), V(2:end, 2:end), curl);
+% Remember that by default, quiver will scale the length of the arrows!
 title('Mimetic-curl');
 xlabel('x')
 ylabel('y')
@@ -61,7 +57,7 @@ set(gcf, 'color', 'w')
 view(0, 90)
 axis tight
 
-% Modified vector field (F*)
+% This P and Q will produce a scalar curl = 2
 function U = P(~, Y)
   U = -Y;
 end
