+++ mole_C++/utils.h
@@ -12,6 +12,7 @@ public:
     static sp_mat spkron(const sp_mat &A, const sp_mat &B);
     static sp_mat spjoin_rows(const sp_mat &A, const sp_mat &B);
     static sp_mat spjoin_cols(const sp_mat &A, const sp_mat &B);
+    static vec spsolve_eigen(const sp_mat &A, const vec &b);
 };
 
 #endif // UTILS_H
