### 2021-07-22 748c5dc
Update elliptic1D.cpp
