+++ mole_C++/Makefile
@@ -2,6 +2,8 @@
 ARMA = /home/johnny/armadillo-10.2.1
 
 # Put here the path to Eigen (optional)
+# You won't need it unless you have to solve a linear system in which
+# case we recommend to use SuperLU over Eigen for the sparse LU factorization
 #EIGEN = /usr/include/eigen3
 
 #----------------------------------------------------------------------
