+++ examples_MATLAB/elliptic3D.m
@@ -11,9 +11,9 @@ n = 6; % -> 8
 o = 7; % -> 9
 
 L = lap3D(k, m, 1, n, 1, o, 1); % 3D Mimetic laplacian operator
-L = L + robinBC3D(k, m, 1, n, 1, o, 1, 1, 0);
+L = L + robinBC3D(k, m, 1, n, 1, o, 1, 1, 0); % Dirichlet BC
 
-RHS = zeros((m+2), (n+2), (o+2));
+RHS = zeros(m+2, n+2, o+2);
 
 RHS(:, :, 1) = 100; % Known value at the cube's front face
 
@@ -21,13 +21,13 @@ RHS = reshape(RHS, (m+2)*(n+2)*(o+2), 1);
 
 SOL = L\RHS;
 
-SOL = reshape(SOL, (m+2), (n+2), (o+2));
+SOL = reshape(SOL, m+2, n+2, o+2);
 
 p = 2; % Page to be displayed
 
 page = SOL(:, :, p);
 
-imagesc(reshape(page, (m+2), (n+2))')
+imagesc(reshape(page, m+2, n+2)')
 title([num2str(p) ' page'])
 xlabel('m')
 ylabel('n')
