### 2021-07-22 098bb27
Create minimal_poisson2D.m
