+++ examples_C++/Makefile
@@ -28,7 +28,7 @@ LIBS += -lsuperlu
 endif
 
 # Rules:
-all: transport1D elliptic1D
+all: transport1D elliptic1D elliptic2D
 
 transport1D: transport1D.cpp
 	$(CXX) $(CXXFLAGS) $(INCPATH) -o transport1D transport1D.cpp $(LIBS)
@@ -36,5 +36,8 @@ transport1D: transport1D.cpp
 elliptic1D: elliptic1D.cpp
 	$(CXX) $(CXXFLAGS) $(INCPATH) -o elliptic1D elliptic1D.cpp $(LIBS)
 
+elliptic2D: elliptic2D.cpp
+	$(CXX) $(CXXFLAGS) $(INCPATH) -o elliptic2D elliptic2D.cpp $(LIBS)
+
 clean:
-	rm -f transport1D elliptic1D
+	rm -f transport1D elliptic1D elliptic2D
