+++ examples_MATLAB/wave2D_case2.m
@@ -29,16 +29,16 @@ I = interpol2D(m, n, 0.5, 0.5);
 I2 = interpolD2D(m, n, 0.5, 0.5);
 
 % Wave propagation speed
-c = 100;
+speed = 100;
 
 % "Force" function
-F = @(x, c) (c^2)*L*x;
+F = @(x, speed) (speed^2)*L*x;
 
 % Simulation time
 TIME = 0.3;
 
 % Temporal discretization based on CFL condition
-dt = dx/(2*c); % dt = h on Young's paper
+dt = dx/(2*speed); % dt = h on Young's paper
 
 % Initial condition
 ICU = @(x, y) sin(pi*x).*sin(pi*y).*(x>2).*(x<3).*(y>2).*(y<3);
@@ -62,7 +62,7 @@ I2 = 0.5*dt*I2;
 for t = 1 : TIME/dt
     % Apply "position Verlet" algorithm -----------------------------------
     uold = uold + I2*vold;
-    vnew = vold + I*F(uold, c);
+    vnew = vold + I*F(uold, speed);
     unew = uold + I2*vnew;
     
     % Update
@@ -80,6 +80,7 @@ for t = 1 : TIME/dt
     set(gcf, 'color', 'w')
     xlabel('x')
     ylabel('y')
+    zlabel('z')
     shading interp
     drawnow
     
