### 2021-07-23 f147077
Update wave2D.m
