### 2021-07-25 7e8a8de
Create test_lap3DCurv_case2.m
