### 2021-07-29 6aeea05
Create sidedNodal.m
Useful for upwind schemes
