+++ mole_MATLAB/sidedNodal.m
@@ -1,0 +1,34 @@
+function S = sidedNodal(k, m, dx, type)
+% Returns a m+1 by m+1 one-dimensional sided approximation for uniformly
+% spaced data points.
+%
+% Parameters:
+%                k : Order of accuracy (either 1 or 2)
+%                m : Number of cells
+%               dx : Step size
+%             type : 'backward' or 'forward'
+
+    if k == 1 % first-order
+        if strcmp(type, 'backward')
+            S = spdiags([-ones(m+1, 1) ones(m+1, 1)], [-1 0], m+1, m+1);
+            S(1, end) = -1;
+        else % forward
+            % S = circshift(backward, -1) or:
+            S = spdiags([-ones(m+1, 1) ones(m+1, 1)], [0 1], m+1, m+1);
+            S(end, 1) = 1;
+        end
+        S = S/dx;
+    else % second-order
+        if strcmp(type, 'backward')
+            S = spdiags([ones(m+1, 1) -4*ones(m+1, 1) 3*ones(m+1, 1)], [-2 -1 0], m+1, m+1);
+            S(1, end-1:end) = [1 -4];
+            S(2, end) = 1;
+        else % forward
+            % S = -rot90(backward, 2) or:
+            S = spdiags([-3*ones(m+1, 1) 4*ones(m+1, 1) -ones(m+1, 1)], [0 1 2], m+1, m+1);
+            S(end-1, 1) = -1;
+            S(end, 1:2) = [4 -1];
+        end
+        S = S/(2*dx);
+    end
+end
