### 2021-07-29 1df3a17
Create sidedNodal2D.m
Useful to compute: Ux, Uy and Uz (individual derivatives) using one-sided approximations
