### 2021-07-30 b5c9c7e
Create sidedNodal3D.m
Useful for Ux, Uy and Uz
