### 2021-07-30 03cf402
Update sidedNodal.m
