### 2021-07-30 d80550a
Update sidedNodal.m
