+++ examples_MATLAB/hyperbolic1D_upwind.m
@@ -16,7 +16,9 @@ dx = (east-west)/m;
 t = 1; % Simulation time
 dt = dx/abs(a); % CFL condition for explicit schemes
 
-S = sidedNodal(1, m, dx, 'backward');
+S = sidedNodal(1, m, dx, 'backward'); % Use 'forward' if a < 0
+% Line 19 can be rewritten in compact form, where based on the sign
+% of (a) either forward or backward differencing is used.
 
 % 1D nodal grid
 grid = west : dx : east;
