+++ mole_C++/Makefile
@@ -1,10 +1,14 @@
+ifndef ARMA
 # Put here the path to Armadillo
 ARMA = /home/johnny/armadillo-10.2.1
+endif
 
+ifndef EIGEN
 # Put here the path to Eigen (optional)
 # You won't need it unless you have to solve a linear system in which
 # case we recommend to use SuperLU over Eigen for the sparse LU factorization
 #EIGEN = /usr/include/eigen3
+endif
 
 #----------------------------------------------------------------------
 # Compiler for C++
