### 2021-08-01 953bc9c
Update sidedNodal.m
