### 2021-08-01 66acbbe
Create hyperbolic1D_lax_friedrichs.m
