### 2021-08-01 53b47a4
Update hyperbolic1D.m
