+++ examples_MATLAB/hyperbolic1D.m
@@ -38,7 +38,6 @@ U2 = U + D/2*U; % Compute one step using Euler's method
 
 % Time integration loop
 for i = 1 : t/dt
-    
     plot(grid, U2, 'o-')
     hold on
     plot(grid, sin(2*pi*(grid - a*i*dt))) % Plot exact solution
@@ -47,7 +46,7 @@ for i = 1 : t/dt
     title(str)
     xlabel('x')
     ylabel('u(x)')
-    ylim([-1 1])
+    axis([west east -1.5 1.5])
     pause(0.04)
     U3 = U + D*U2; % Compute next step using Leapfrog scheme
     U = U2;
