### 2021-08-01 8693a7c
Update hyperbolic1D_lax_friedrichs.m
