### 2021-08-01 a4ecf21
Update hyperbolic1D_lax_friedrichs.m
