### 2021-08-01 93094ac
Update hyperbolic1D_lax_friedrichs.m
