### 2021-08-01 b9a02c7
Update hyperbolic1D.m
