### 2021-08-11 jcorbino/mole#8 @jshipton
Installation instructions
[https://github.com/openjournals/joss-reviews/issues/3380](url) It would be great to have installation instructions for Mac users e.g. `brew install superlu`. I'd add them myself but I'm trying to avoid upgrading my OS as I'm going to buy a new laptop soon so anything I do may well be out of date!
