### 2021-08-11 jcorbino/mole#8 @jcorbino
As a matter of fact, I have done it, but it was mainly to test the parallel LU decomposition via SuperLU_Dist in a couple of problems. That I remember, I pretty much followed the same steps EXCEPT that since I didn't have root privileges, I had to locally build and link against SuperLU. What I can do is to add the instructions for Mac users between today and tomorrow (since I'm a bit busy today) so you can try it on your laptop, would that work for you @jshipton?
