### 2021-08-12 jcorbino/mole#8 @jshipton
That sounds like a good plan! I have tried a few things myself but have not made it work yet. I'm afraid I'm on holiday for the next 2 weeks so I will complete as much of the rest of the review as I can but I'm not sure how far I'll get with installing...
