### 2021-08-26 27e313d
Update burgers1D.m
