+++ README.md
@@ -13,7 +13,7 @@ on staggered grids (uniform and nonuniform) and they satisfy local and
 global conservation laws.
 
 The mathematics is based on the work of [Corbino and Castillo, 2020]. 
-However the user may find useful previous publications such as [Castillo and Grone, 2003],
+However, the user may find useful previous publications such as [Castillo and Grone, 2003],
 in which similar operators are derived using a matrix analysis approach.
 
 
