### 2021-08-30 1b1353b
Update hyperbolic1D.m
