### 2021-08-30 40e80df
Update hyperbolic1D_lax_friedrichs.m
