### 2021-08-30 fed5f9f
Update burgers1D.m
