### 2021-09-07 #3380 @jshipton
Apologies for the slow reply - I've been on leave and am now at a conference. I was in touch with the author (see https://github.com/jcorbino/mole/issues/8) about installation instructions for Macs...
