### 2021-09-07 jcorbino/mole#8 @jshipton
Hi @jcorbino! Any progress on the Mac installation instructions?
