### 2021-10-06 jcorbino/mole#8 @jcorbino
Hi @jshipton I'm so sorry for the delay with the instructions for Mac users. I just wanted to let you know that tomorrow (Wednesday, October 6th) I'll push the updated "README" file with a section explaining the steps for MacOS. Basically the best way to go is by installing Eigen instead of SuperLU. But of course, there are a couple of flags that must be changed in MOLE'S Makefile for it to work on a Mac. Tomorrow I'll let you know.

Thank you,
