+++ mole_MATLAB/vanderCoef.m
@@ -1,0 +1,4 @@
+function C = vanderCoef(x)
+    V = inv(fliplr(vander(x)));
+    C = V(1, :); % Coefficients for the interiors
+end
