### 2021-10-07 jcorbino/mole#8 @jcorbino
Here you go @jshipton

https://www.dropbox.com/s/0u7x9rrqy1onvst/MOLE%20installation%20on%20MacOS.pdf?dl=0

Thanks,
