### 2021-10-07 #3380 @jcorbino
Here's a one page document with steps to install MOLE on MacOS. [https://www.dropbox.com/s/0u7x9rrqy1onvst/MOLE%20installation%20on%20MacOS.pdf?dl=0](url) . Basically, for the average Mac user, the way to go is by using Eigen and not SuperLU. The document has also been uploaded to the root directory of the repo.

Kind regards,
