### 2021-10-19 d7fbef4
Create nodal2DCurv.m
