### 2021-10-19 d510e1e
Create nodal3DCurv.m
