+++ mole_MATLAB/nodal3DCurv.m
@@ -1,0 +1,32 @@
+function [Nx, Ny, Nz] = nodal3DCurv(k, X, Y, Z)
+    % Get the determinant of the jacobian and the metrics
+    [J, Xe, Xn, Xc, Ye, Yn, Yc, Ze, Zn, Zc] = jacobian3D(k, X, Y, Z);
+    
+    % Dimensions of nodal grid
+    [n, m, o] = size(X);
+    
+    len = n*m*o;
+    
+    % Convert metrics to diagonal matrices
+    J = spdiags(1./J, 0, numel(len), numel(len));
+    A = spdiags(Yn.*Zc-Zn.*Yc, 0, len, len);
+    B = spdiags(Zn.*Xc-Xn.*Zc, 0, len, len);
+    C = spdiags(Xn.*Yc-Yn.*Xc, 0, len, len);
+    D = spdiags(Ze.*Yc-Ye.*Zc, 0, len, len);
+    E = spdiags(Xe.*Zc-Ze.*Xc, 0, len, len);
+    F = spdiags(Ye.*Xc-Xe.*Yc, 0, len, len);
+    G = spdiags(Ye.*Zn-Ze.*Yn, 0, len, len);
+    H = spdiags(Ze.*Xn-Xe.*Zn, 0, len, len);
+    I = spdiags(Xe.*Yn-Ye.*Xn, 0, len, len);
+    
+    % Construct 3D uniform nodal operator
+    N = nodal3D(k, m, 1, n, 1, o, 1); % N is tall and skinny
+    Ne = N(1:len, :);
+    Nn = N(len+1:2*len, :);
+    Nc = N(2*len+1:end, :);
+    
+    % Apply transformation
+    Nx = J*(A*Ne+D*Nn+G*Nc);
+    Ny = J*(B*Ne+E*Nn+H*Nc);
+    Nz = J*(C*Ne+F*Nn+I*Nc);
+end
