### 2021-10-19 716ac0f
Update nodal3DCurv.m
