+++ mole_MATLAB/boundaryIdx2D.m
@@ -6,15 +6,15 @@ function idx = boundaryIdx2D(m, n)
 %           m : Number of nodes along x-axis
 %           n : Number of nodes along y-axis
 
-    idx = zeros(m*n, 1);
+    idx = zeros(2*m+2*(n-2), 1);
     
-    idx(1:m) = 1;
-    idx(end-m+1:end) = 1;
+    idx(1:m) = 1:m;
+    idx(end-m+1:end) = m*n-m+1:m*n;
     
-    for i = m+1:m:m*n
-        idx(i) = 1;
-        idx(i+m-1) = 1;
+    k = m+1;
+    for i = m+1:m:m*n-m
+        idx(k) = i;
+        idx(k+1) = i+m-1;
+        k = k+2;
     end
-    
-    idx = find(idx);
 end
