### 2021-10-20 da27031
Create elliptic2D_nodal_curv_sinusoidal.m
