### 2021-10-20 4b804eb
Update elliptic2D_nodal_curv_sinusoidal.m
