+++ JOSS/paper.md
@@ -25,15 +25,15 @@ MOLE is a high quality (C++ & MATLAB) library that implements high-order mimetic
 
 The mathematics is based on the work of [@Corbino]. However the user may find useful previous publications such as [@Castillo], in which similar operators are derived using a matrix analysis approach.
 
-# Statement of need
-
-Implementing mimetic operators is not a trivial matter, particularly in three dimensions,  this is substantially facilitated by MOLE relieving the user to devote their time to focus on the problem of interest. The user interested in solving, for example, a Poisson equation  -Laplacian u = f, will be solving a discrete analog of this equation, -DGu = f, by using MOLE and a few lines of code.
-
 # Mimetic operators
 
 All linear transformations can be represented by a matrix multiplication, integration and differentiation are linear transformations. Mimetic operators are essentially matrices that when applied to discrete scalar or vector fields produce high-order approximations that are faithful to the physics.
 
 The basis of higher-dimensional operators, as well of more sophisticated operators such as the Laplacian or the Biharmonic operator are the one-dimensional mimetic Gradient (**G**) and Divergence (**D**) operators. These operators can be reused throughout the model and they provide a higher level of abstraction at the time of solving differential equations.
+
+# Statement of need
+
+Implementing mimetic operators is not a trivial matter, particularly in three dimensions,  this is substantially facilitated by MOLE relieving the user to devote their time to focus on the problem of interest. The user interested in solving, for example, a Poisson equation  -Laplacian u = f, will be solving a discrete analog of this equation, -DGu = f, by using MOLE and a few lines of code.
 
 # The library
 
