+++ JOSS/paper.md
@@ -35,6 +35,10 @@ The basis of higher-dimensional operators, as well of more sophisticated operato
 
 Implementing mimetic operators is not a trivial matter, particularly in three dimensions,  this is substantially facilitated by MOLE relieving the user to devote their time to focus on the problem of interest. The user interested in solving, for example, a Poisson equation  -Laplacian u = f, will be solving a discrete analog of this equation, -DGu = f, by using MOLE and a few lines of code.
 
+# State of the field
+
+A previous library [@mtk] was developed to implement the mimetic operators presented in the Castillo and Grone paper [@Castillo].  This library was only capable of handling dense matrices so it was limited to solve small problems hence its development was stopped. MOLE implements the operators presented in the Corbino and Castillo paper [@Corbino]. These  operators are optimal from the number of points in each stencil and produce more accurate results. MOLE deals with sparse matrices efficiently and is capable of solving problems with millions of cells.
+
 # The library
 
 MOLE was designed to be an intuitive software package to numerically solve partial differential equations using mimetic methods. MOLE is implemented in C++ and in MATLAB scripting language (these are two independent flavors) and every single function in MOLE returns a sparse matrix of the requested mimetic operator. For information on the installation or usage of the library, please read the [User's Manual](https://github.com/jcorbino/mole/blob/master/MOLE%20User's%20Manual%20(MATLAB%20version).pdf) which is included in the repository.
