+++ JOSS/paper.bib
@@ -30,3 +30,14 @@
   	year = 2016,
     volume = 1
 }
+
+@article{Mtk,
+  	url = {https://www.sciencedirect.com/science/article/pii/S037704271300719X?via%3Dihub},
+  	author = {{Sanchez} Eduardo J. and {Castillo}, J. E.},
+  	doi = {10.1016/j.cam.2013.12.046},
+  	journal = {Computational and Applied Mathematics},
+  	month = November,
+  	title = {{The Mimetic Methods Toolkit: An object-oriented API for Mimetic Finite Differences}},
+  	year = 2014,
+    volume = 270
+}
