+++ JOSS/paper.bib
@@ -33,7 +33,7 @@
 
 @article{Mtk,
   	url = {https://www.sciencedirect.com/science/article/pii/S037704271300719X?via%3Dihub},
-  	author = {{Sanchez} Eduardo J. and {Castillo}, J. E.},
+  	author = {{Sanchez}, Eduardo J. and {Castillo}, J. E.},
   	doi = {10.1016/j.cam.2013.12.046},
   	journal = {Computational and Applied Mathematics},
   	month = November,
