### 2021-11-04 #3380 @jcorbino
Hello @jshipton and @ketch,

Did you have a chance to read the last version of the draft? I think I addressed all your comments. @jshipton please refer to https://github.com/jcorbino/mole/blob/master/MOLE%20installation%20on%20MacOS.pdf for instructions on installation on Macs.

Kind regards,
