### 2021-11-05 #3380 @jedbrown
Thanks for the updates @jcorbino. Would you consider it appropriate to mention connections with mixed finite element methods (algebraically equivalent in some limited cases, and able to preserve similar properties)?

Can MOLE be used for problems like shallow water with Coriolis or the p-Laplacian? Also, I feel the paper misses an opportunity to demonstrate the benefit of mimetic methods by showing how standard methods may fail to converge on certain problems or do not preserve key properties of the continuum solutions. I know the mimetic literature discusses these things, but you may have readers who are unfamiliar with those examples and solving a 1D Poisson problem in the viewgraph norm is ... underwhelming.

I notice that your [users manual](https://github.com/jcorbino/mole/blob/master/MOLE%20User's%20Manual%20(MATLAB%20version).pdf) says

> MOLE is distributed under a "dual-license" model. GPLv3 for academic and nonprofit purposes, and copyright for commercial intentions.

This is not acceptable as written. GPLv3 absolutely can be used commercially (that's Red Hat's business, for example) and you can't imply that it's only GPLv3 for non-commercial purposes. It isn't open source if a company can't sell support for an application built on it, or use in an in-house product. You may advertise that alternate licensing is available if a user isn't satisfied with the terms of GPLv3, though this will have an impact on your contributing guidelines. (Would a valued contributor get paid as part of such licensing? Must they sign a copyright assignment or grant you a different license when submitting their PR so that you have authority to grant a commercial user a different license? Note that inbound=outbound is the de facto assumption in open source contributions, sometimes codified through the likes of a Developers Certificate of Origin or a Contributor License Agreement.)

Less importantly, the OSX installation docs are basically text in a PDF without source, which makes it hard to edit (e.g., for a contributor who streamlines something or needs to make a note for a newer OS version). Those would be easier to read and copy/paste from if converted to markdown (a 2-minute copy/paste job).
