### 2021-11-07 #3380 @ketch
- The section "Mimetic Operators" does not, as I understand it, actually state what a mimetic operator is.  "faithful to the physics" is too vague.  I believe "mimetic" has a mathematical meaning as it is used in the literature (by the authors!) and that meaning should be stated here.
- The 3rd reference seems to be missing an author.  Please check the whole bibliography carefully.
- The "state of the field" only mentions one other library, written by some of the authors of this manuscript.  But there are a vast number of other libraries out there that aim to "produce high-order approximations that are faithful to the physics".  There's no indication of "how this software compares to other commonly-used packages" (quoting the review checklist above).
- The last point from my earlier list has not been addressed in the paper.
- The writing quality (English usage) is not very high.  If desired, I can provide a list of examples.

I agree with @jedbrown that using Laplace as your main example seems to totally miss the point of mimetic methods.
