+++ JOSS/paper.bib
@@ -31,13 +31,12 @@
     volume = 1
 }
 
-@article{Mtk,
-  	url = {https://www.sciencedirect.com/science/article/pii/S037704271300719X?via%3Dihub},
-  	author = {{Sanchez}, Eduardo and {Castillo}, J. E.},
-  	doi = {10.1016/j.cam.2013.12.046},
-  	journal = {Computational and Applied Mathematics},
-  	month = November,
-  	title = {{The Mimetic Methods Toolkit: An object-oriented API for Mimetic Finite Differences}},
-  	year = 2014,
-    volume = 270
+@article{MTK,
+  title={The mimetic methods toolkit: an object-oriented api for mimetic finite differences},
+  author={Sanchez, Eduardo J and Paolini, Christopher P and Castillo, Jose E},
+  journal={Journal of Computational and Applied Mathematics},
+  volume={270},
+  pages={308--322},
+  year={2014},
+  publisher={Elsevier}
 }
