+++ JOSS/paper.bib
@@ -9,6 +9,13 @@
     volume = 364
 }
 
+@book{Miranda,
+  title={Mimetic discretization methods},
+  author={Castillo, Jos{\'e} E and Miranda, Guillermo F},
+  year={2013},
+  publisher={CRC Press}
+}
+
 @article{Castillo,
   	url = {https://doi.org/10.1137/S0895479801398025},
   	author = {{Castillo}, J. E. and {Grone}, R. D.},
