### 2021-11-08 #3380 @jcorbino
@jedbrown @ketch thank you for the feedback. **I think I've addressed all of them in the latest version of the draft** (which was generated right before this comment).

* License updated to GPLv3.

* The one-dimensional Poisson problem was chosen for conciseness purposes. The point is that even such an "underwhelming" problem will require considerably more lines of code if one had chosen a different technique. In the new version of the draft, I've expressed that our library comes with over 30 examples (from simple linear and one-dimensional problems, to three-dimensional and nonlinear problems, steady-state and time-dependent).

* I want to emphasize that this article is only about the library and not about how mimetic methods compare to other alternatives such as finite-element based methods, that's a different article.

* To the best of my knowledge, there are no other libraries that implement mimetic methods (at least not public).

@ketch I would greatly appreciate if you share your list indicating the shortcomings in the writing of the article, I have to admit that your comment caught me by surprise. Lastly, thank you for pointing out that one of the references was missing an author! (fixed)
