### 2021-11-08 22266b7
Delete LICENSE.txt
