+++ LICENSE.txt
@@ -1,18 +1,0 @@
-MOLE: Mimetic Operators Library Enhanced
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-The GNU General Public License is a copyleft license, meaning any
-modified version must be licensed under the GNU GPL as well, including
-any new code written.
