### 2021-11-08 #3380 @jcorbino
Also in the new draft:  _"It is important to mention that MOLE's main role is the construction of matrices that represent spatial derivative operators and boundary conditions; other components such as solvers and time steppers are only provided via self-contained examples."_ I hope this is clear enough.
