### 2021-11-08 #3380 @jcorbino
@jedbrown @ketch @jshipton I could also remove the example (since it is underwhelming) and not show any just like in these six recently accepted JOSS papers:

- excluder: An R package that checks for exclusion criteria in online data
- grlc: the git repository linked data API constructor
- spant: An R package for magnetic resonance spectroscopy analysis
- CategoricalTimeSeries.jl: A toolbox for categorical time-series analysis
- Qlunc: Quantification of lidar uncertainty
- HyRiver: Hydroclimate Data Retriever

It should be acceptable.
