### 2021-11-13 #3380 @jedbrown
There's a sizable intersection of mixed FE and mimetic FD methods that are algebraically equivalent. FEEC gives a range of methods that satisfy the discrete compatibility, albeit going to high order in a different way (see ["periodic table" of finite elements](https://www-users.cse.umn.edu/~arnold/femtable/index.html)). As example of broader impact, note that FEEC is the foundation of the UK Met Office's new dynamical core (will be used for weather and climate). I'm not an expert in SBP methods, but I understand they also have some overlap. This is a JOSS paper on one such toolkit: https://joss.theoj.org/papers/10.21105/joss.03454.pdf

I tried running `convection_diffusion` and get an error
```
error: interp3: X, Y, Z, and V dimensions must be equal
```
I'm using Octave, but this isn't the usual kind of error one gets when Octave doesn't implement something needed by a Matlab script.

I don't see how to run tests. There isn't a `make test` much less continuous integration. I understand that someone can go into the examples and run them, but it's unclear how to determine whether it's correct. I understand that some produce figures that show analytic and numerical solutions, but they only plot the solution and don't do a grid convergence study. I also see that hyperbolic examples use forward Euler or Leapfrog with a time step of exactly 1 CFL on a uniform grid, which isn't discerning (e.g., first order upwind is exact). If you compare SummationByPartsOperators.jl (linked above), you'll see that there's a github action and `runtests.jl` (the Julia convention) that runs timed tests including comparisons of conventional and SBP operators. All these tests return clear pass/fail so that subjective or expert user knowledge is not needed to interpret the results. As numerical analysts, we should take verification seriously enough to put in our repositories sufficient tooling to test order of accuracy and other discretization properties that we claim to provide.

Honestly, I would drop the phrase "faithful to the physics" -- these discretizations don't preserve positivity for advection, for example. There's good reason for this (Godunov's theorem), but I would just state what is preserved (discrete de Rham complex). I think it's worth writing out the fundamental properties that define mimetic methods rather than outsourcing it to a paywalled paper. It's also important to discuss limitations. I don't know if MOLE would be appropriate for nonlinear spatial discretizations. The `elliptic1D` example produces a nonsymmetric matrix that is indefinite and has complex eigenvalues -- this might not be appropriate if one wishes to do eigenanalysis. Our documentation should clearly state what is available, what is in-scope but not yet implemented, and what is out of scope. A prospective user shouldn't need to become an expert in MFD to determine whether MOLE will be suitable for their application.

The `doc_MATLAB` docs work in Chromium, but not in Firefox. Some of the docstrings are wrong. For example, `div2DCurv` says "Get the determinant of the jacobian and the metrics", but it returns a divergence operator. Am I reading correctly that curl is only supported for uniform grids in 2D? Maybe a table showing operator, dimension, and uniform, non-uniform Cartesian, and curvilinear would be helpful. This could go in the project docs (maybe README) and/or in the paper.
