+++ JOSS/paper.bib
@@ -47,3 +47,34 @@
   year={2014},
   publisher={Elsevier}
 }
+
+@article{Bazan,
+  title={Mimetic finite difference methods in image processing},
+  author={Bazan, Carlos and Abouali, M and Castillo, J and Blomgren, P},
+  journal={Computational \& Applied Mathematics},
+  volume={30},
+  number={3},
+  pages={701--720},
+  year={2011},
+  publisher={SciELO Brasil}
+}
+
+@article{Boada,
+  title={High-order mimetic finite differences for anisotropic elliptic equations},
+  author={Boada, Angel and Paolini, Christopher and Castillo, Jose E},
+  journal={Computers \& Fluids},
+  volume={213},
+  pages={104746},
+  year={2020},
+  publisher={Elsevier}
+}
+
+@article{Boada2,
+  title={High order mimetic difference simulation of unsaturated flow using Richards equation},
+  author={Velazco, Angel Boada and Corbino, Johnny and Castillo, Jose},
+  journal={Mathematics in Applied Sciences and Engineering},
+  volume={1},
+  number={4},
+  pages={401--409},
+  year={2020}
+}
