+++ JOSS/paper.md
@@ -21,15 +21,17 @@ bibliography: paper.bib
 
 # Summary
 
-MOLE is a high quality (C++ & MATLAB) library that implements high-order mimetic operators to solve partial differential equations. It provides discrete analogs of the most common vector calculus operators: Gradient, Divergence, Laplacian and Curl. These operators (matrices) act on staggered grids (uniform and nonuniform) and they satisfy local and global conservation laws.
+MOLE is a high quality (C++ & MATLAB) library that implements high-order mimetic operators. It provides discrete analogs of the most common vector calculus operators: Gradient, Divergence, Laplacian and Curl. These operators (matrices) act on staggered grids (uniform and nonuniform) and they satisfy local and global conservation laws. These operators can be used to build codes to solve partial differential equations (PDEs).
 
-The mathematics is based on the work of [@Corbino]. However, the user may find useful previous publications such as [@Castillo], in which similar operators are derived using a matrix analysis approach.
+The mathematics is based on the work of [@Corbino]. In addition, the user may find useful previous publications such as [@Castillo], in which similar operators are derived using a matrix analysis approach.
 
 # Mimetic operators
 
-All linear transformations can be represented by a matrix multiplication, integration and differentiation are linear transformations. Mimetic operators are essentially matrices that when applied to discrete scalar or vector fields produce high-order approximations that are faithful to the physics. Specifically, our operators are uniformly accurate, and they satisfy the _discrete extended Gauss theorem_ [@Miranda].
+Mimetic finite-difference operators, divergence (**D**), gradient (**G**), curl (**C**) and laplacian (**L**) are discrete analogs of their corresponding continuum operators. These mimetic finite-difference operators satisfy in the discrete sense the vector identities that the continuum ones do, making them more faithful to the physics.
 
-The basis of higher-dimensional operators, as well of more sophisticated operators such as the Laplacian or the Biharmonic operator are the one-dimensional mimetic Gradient (**G**) and Divergence (**D**) operators. These finite-dimensional operators can be reused throughout the model and they provide a higher level of abstraction at the time of solving differential equations.
+The basis of higher-dimensional operators, as well of more sophisticated operators such as the laplacian or the biharmonic operator are the one-dimensional mimetic **G** and **D** operators. These finite-dimensional operators can be reused throughout the model and they provide a higher level of abstraction at the time of solving differential equations.
+
+These operators, have been used to write codes to solve PDEs of different types [@Bazan, @Boada, Boada2]. For an overview of mimetic methods of different types see the book by Castillo and Miranda and the references there in [@Miranda].
 
 # Statement of need
 
@@ -43,7 +45,7 @@ A previous library [@MTK] was developed to implement the mimetic operators prese
 
 MOLE was designed to be an intuitive software package to numerically solve partial differential equations using mimetic methods. MOLE is implemented in C++ and in MATLAB scripting language (these are two independent flavors) and every single function in MOLE returns a sparse matrix of the requested mimetic operator. For information on the installation or usage of the library, please read the [User's Manual](https://github.com/jcorbino/mole/blob/master/MOLE%20User's%20Manual%20(MATLAB%20version).pdf) which is included in the repository.
 
-For example, if the user wants to get a one-dimensional *k*-order mimetic Laplacian, just need to invoke:
+Mimetic operators can be easily used to build codes to solve PDEs with a few lines of code. For example, if the user wants to get a one-dimensional *k*-order mimetic Laplacian, just need to invoke:
 ```matlab
 lap(k, m, dx);
 ```
