+++ JOSS/paper.bib
@@ -9,6 +9,47 @@
     volume = 364
 }
 
+@article{Jorge,
+  title={Mimetic finite difference methods for restoration of fundus images for automatic detection of glaucoma suspects},
+  author={Villamizar, Jorge and Calder{\'o}n, Giovanni and Carrillo, Julio and Bautista Rozo, Lola and Carrillo, Juan and Rueda, Juan and Castillo, Jos{\'e}},
+  journal={Computer Methods in Biomechanics and Biomedical Engineering: Imaging \& Visualization},
+  pages={1--8},
+  year={2021},
+  publisher={Taylor \& Francis}
+}
+
+@article{Abouali,
+  title={Stability and Performance Analysis of the Castillo-Grone Mimetic Operators in Conjunction with RK3 Time Discretization in Solving Advective Equations},
+  author={Abouali, Mohammad and Castillo, Jose E},
+  journal={Procedia Computer Science},
+  volume={18},
+  pages={465--472},
+  year={2013},
+  publisher={Elsevier}
+}
+
+@article{Josep,
+  title={Mimetic seismic wave modeling including topography on deformed staggered grids},
+  author={de la Puente, Josep and Ferrer, Miguel and Hanzich, Mauricio and Castillo, Jos{\'e} E and Cela, Jos{\'e} M},
+  journal={Geophysics},
+  volume={79},
+  number={3},
+  pages={T125--T141},
+  year={2014},
+  publisher={Society of Exploration Geophysicists}
+}
+
+@article{Rojas,
+  title={Modelling of rupture propagation using high-order mimetic finite differences},
+  author={Rojas, Otilio and Day, Steven and Castillo, Jose and Dalguer, Luis A},
+  journal={Geophysical Journal International},
+  volume={172},
+  number={2},
+  pages={631--650},
+  year={2008},
+  publisher={Blackwell Publishing Ltd Oxford, UK}
+}
+
 @book{Miranda,
   title={Mimetic discretization methods},
   author={Castillo, Jos{\'e} E and Miranda, Guillermo F},
