+++ mole_MATLAB/div2DCurv.m
@@ -1,5 +1,5 @@
 function D = div2DCurv(k, X, Y)
-% Returns a 2D curvilinear mimetic gradient
+% Returns a 2D curvilinear mimetic divergence
 
     % Get the determinant of the jacobian and the metrics
     [J, Xe, Xn, Ye, Yn] = jacobian2D(k, X, Y);
