### 2021-11-29 b1f2ce7
Update grad2DCurv.m
