### 2021-11-29 b12aa96
Update div3DCurv.m
