### 2021-11-29 14da93e
Update grad3DCurv.m
