### 2021-12-01 f429c88
Update neumann3DCurv.m
