### 2021-12-17 dcfa925
Update lock_exchange.m
