### 2021-12-19 4f2d0c9
Update utils.h
