### 2021-12-31 5502091
Update burgers1D.m
