### 2022-02-13 af9da08
Create spsolve_cuda.cu
This isolated example illustrates how to solve a linear system where 'A' is of type sp_mat and 'b' is of type vec (Armadillo's types) using sparse QR factorization provided by cuSOLVER.
