+++ examples_C++/spsolve_cuda.cu
@@ -1,0 +1,176 @@
+// Compile:
+// nvcc spsolve_cuda.cu -o spsolve_cuda -O3 -I./armadillo-10.2.1/include -lcusparse -lcusolver
+
+#include <cusolverSp.h>
+#define ARMA_ALLOW_FAKE_GCC
+#include <armadillo>
+
+using namespace arma;
+
+double* cuda_qr(const double* values,
+        const long long unsigned int a_nnz,
+        const long long unsigned int rows,
+        const long long unsigned int cols,
+        const long long unsigned int* row_ind,
+        const long long unsigned int* col_ptrs,
+        double* b,
+        double* x) {
+
+  int nnz = (int)a_nnz;
+  int* h_csccol_pts;
+  int* h_cscRowInd;
+  double* h_cscVal;
+
+  h_csccol_pts = (int*)malloc(sizeof(int) * (cols + 1));
+  for (int i = 0; i < cols + 1; ++i)
+    h_csccol_pts[i] = (int)col_ptrs[i];
+
+  h_cscRowInd = (int*)malloc(sizeof(int) * nnz);
+  for (int i = 0; i < nnz; ++i)
+    h_cscRowInd[i] = (int)row_ind[i];
+
+  h_cscVal = (double*)malloc(sizeof(double) * nnz);
+  memcpy(h_cscVal, values, sizeof(double) * nnz);
+
+  size_t h_buffer = 0;
+  double* d_buffer = nullptr;
+
+  int* d_csccol_pts = nullptr;
+  int* d_cscRowInd = nullptr;
+  int* d_csrRowPtr = nullptr;
+  int* d_csrColInd = nullptr;
+  double* d_csrvalues = nullptr;
+  double* d_cscVal = nullptr;
+
+  cusparseHandle_t sphandle = nullptr;
+  cusparseCreate(&sphandle);
+
+  cudaMalloc((void**)&d_cscVal, sizeof(double) * nnz);
+  cudaMalloc((void**)&d_csccol_pts, sizeof(int) * (cols + 1));
+  cudaMalloc((void**)&d_cscRowInd, sizeof(int) * nnz);
+  cudaMalloc((void**)&d_csrvalues, sizeof(double) * nnz);
+  cudaMalloc((void**)&d_csrRowPtr, sizeof(int) * (rows + 1));
+  cudaMalloc((void**)&d_csrColInd, sizeof(int) * nnz);
+
+  cusparseIndexBase_t idxBase = CUSPARSE_INDEX_BASE_ZERO;
+  cusparseAction_t copyValues = CUSPARSE_ACTION_NUMERIC;
+  cusparseCsr2CscAlg_t alg = CUSPARSE_CSR2CSC_ALG2;
+  cudaDataType valType = CUDA_R_64F;
+
+  cudaMemcpy(d_cscVal, h_cscVal, sizeof(double) * nnz, cudaMemcpyHostToDevice);
+  cudaMemcpy(d_csccol_pts, h_csccol_pts, sizeof(int) * (cols + 1), cudaMemcpyHostToDevice);
+  cudaMemcpy(d_cscRowInd, h_cscRowInd, sizeof(int) * nnz, cudaMemcpyHostToDevice);
+
+  cusparseMatDescr_t descrA = nullptr;
+  cusparseCreateMatDescr(&descrA);
+  cusparseSetMatType(descrA, CUSPARSE_MATRIX_TYPE_GENERAL);
+  cusparseSetMatIndexBase(descrA, CUSPARSE_INDEX_BASE_ZERO);
+
+  cusparseCsr2cscEx2_bufferSize(sphandle,
+                                cols,
+                                rows,
+                                nnz,
+                                d_cscVal,
+                                d_csccol_pts,
+                                d_cscRowInd,
[…]
