+++ tests_C++/test1.cpp
@@ -9,7 +9,7 @@ int main() {
     int k = 2;
     int m = 2*k+1;
     float dx = 1;
-    float tol = 1e-16;
+    float tol = 1e-12;
     
     Divergence D(k, m, dx);
     vec field(m+1, fill::ones);
