+++ tests_C++/test2.cpp
@@ -9,7 +9,7 @@ int main() {
     int k = 2;
     int m = 2*k;
     float dx = 1;
-    float tol = 1e-16;
+    float tol = 1e-12;
     
     Gradient G(k, m, dx);
     vec field(m+2, fill::ones);
